third_party/blink/web_tests/external/wpt/css/css-inline/baseline-source/baseline-source-first-textarea-002.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] .target > * 9
  assert_equals: \n<span data-offset-x="25"></span>\noffsetLeft expected 25 but got 10
[FAIL] .target > * 11
  assert_equals: \n<span data-offset-x="25"></span>\noffsetLeft expected 25 but got 10
Harness: the test ran to completion.
